refactor(inventory): isolate action transaction pipelines
Split acquisition, dismantle, socket, item-state, and web-service actions out of the monolithic state and protocol runtimes. Represent service mutations as an exclusive transaction variant so staging and commit cannot accidentally publish overlapping actions.

Move Family-4 inventory updates into their own snapshot unit, centralize full-family refresh staging, and resync other authenticated peers after a shared account mutation so Family 4, appearance, and roster views do not remain stale.

Replace fixed dismantle material constants with a validated version-4 settings policy, keeping payout definitions and quantities configuration-driven across build variants.
