+++ Sunrise/src/core/settings/settings.h
@@ -18,7 +18,7 @@ namespace sunrise::core::settings {
  * Raise it when a key is renamed, removed, or changes meaning. Adding a key needs no raise,
  * because a missing key already takes its default. An older file is upgraded in place at load.
  */
-inline constexpr std::uint32_t kSettingsVersion = 3;
+inline constexpr std::uint32_t kSettingsVersion = 4;
 
 /** Parsed read-only process settings. */
 struct Settings {
