+++ Sunrise/src/server/bap/encrypted/transactions/service_outcome_commit.cpp
@@ -16,85 +16,71 @@ namespace sunrise::server::bap::encrypted::transactions {
  */
 bool commit(ServiceOutcome& outcome, Publication& publication) noexcept {
     publication = {};
-    const unsigned mutationCount = static_cast<unsigned>(outcome.hasActivitySessionAllocation)
-                                   + static_cast<unsigned>(outcome.hasActivityTransaction)
-                                   + static_cast<unsigned>(outcome.hasMatchmakingMutation)
-                                   + static_cast<unsigned>(outcome.hasEquipmentSwap)
-                                   + static_cast<unsigned>(outcome.hasSocketPlug)
-                                   + static_cast<unsigned>(outcome.hasItemState)
-                                   + static_cast<unsigned>(outcome.hasItemAcquisition)
-                                   + static_cast<unsigned>(outcome.hasProfileItemAcquisition)
-                                   + static_cast<unsigned>(outcome.hasItemDismantle);
-    // A service route may never combine independently versioned State transactions.
-    if (mutationCount > 1U) {
-        return false;
-    }
-    if (outcome.hasActivitySessionAllocation) {
-        const std::uint64_t sessionId = outcome.activitySessionAllocation.sessionId;
+    if (auto* allocation = transaction_if<state::activity::PendingAllocation>(outcome)) {
+        const std::uint64_t sessionId = allocation->sessionId;
         if (sessionId == state::activity::kAbsentSessionId
-            || !state::activity::commit(outcome.activitySessionAllocation)) {
+            || !state::activity::commit(*allocation)) {
             return false;
         }
         publication.activitySessionId = sessionId;
         publication.hasActivitySessionBinding = true;
         return true;
     }
-    if (outcome.hasActivityTransaction) {
-        if (outcome.activityPlan.mutationDomain == activity_message::MutationDomain::entitySlots) {
-            return state::activity::entity_slots::commit(outcome.activityPlan.entitySlotMutation);
+    if (auto* plan = transaction_if<activity_message::ActivityPlan>(outcome)) {
+        if (plan->mutationDomain == activity_message::MutationDomain::entitySlots) {
+            return state::activity::entity_slots::commit(plan->entitySlotMutation);
         }
-        if (outcome.activityPlan.mutationDomain == activity_message::MutationDomain::membership) {
-            return state::activity::membership::commit(outcome.activityPlan.membershipMutation);
+        if (plan->mutationDomain == activity_message::MutationDomain::membership) {
+            return state::activity::membership::commit(plan->membershipMutation);
         }
         // The retained patch epoch is connection state, so it commits nothing here.
-        return outcome.activityPlan.mutationDomain == activity_message::MutationDomain::patchEpoch;
+        return plan->mutationDomain == activity_message::MutationDomain::patchEpoch;
     }
-    if (outcome.hasMatchmakingMutation) {
-        return state::matchmaking::commit(outcome.matchmakingMutation);
+    if (auto* mutation = transaction_if<state::matchmaking::PendingMutation>(outcome)) {
+        return state::matchmaking::commit(*mutation);
     }
-    if (outcome.hasEquipmentSwap) {
-        const bool committed = state::commit_equipment_swap(outcome.equipmentSwap);
+    if (auto* transaction = transaction_if<EquipmentSwapTransaction>(outcome)) {
+        const bool committed = state::commit_equipment_swap(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=equip stage=transaction_commit result=ok"
                                    : "ev=equip stage=transaction_commit result=fail");
         return committed;
     }
-    if (outcome.hasItemAcquisition) {
-        const bool committed = state::commit_item_acquisition(outcome.itemAcquisition);
+    if (auto* transaction = transaction_if<ItemAcquisitionTransaction>(outcome)) {
+        const bool committed = state::commit_item_acquisition(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=acquire stage=transaction_commit result=ok"
                                    : "ev=acquire stage=transaction_commit result=fail");
         return committed;
     }
-    if (outcome.hasSocketPlug) {
-        const bool committed = state::commit_socket_plug(outcome.socketPlug);
+    if (auto* transaction = transaction_if<SocketPlugTransaction>(outcome)) {
+        const bool committed = state::commit_socket_plug(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=socket_plug stage=transaction_commit result=ok"
                                    : "ev=socket_plug stage=transaction_commit result=fail");
         return committed;
     }
-    if (outcome.hasItemState) {
-        const bool committed = state::commit_item_state(outcome.itemState);
+    if (auto* transaction = transaction_if<ItemStateTransaction>(outcome)) {
+        const bool committed = state::commit_item_state(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=item_state stage=transaction_commit result=ok"
                                    : "ev=item_state stage=transaction_commit result=fail");
         return committed;
     }
-    if (outcome.hasProfileItemAcquisition) {
-        const bool committed =
-            state::commit_profile_item_acquisition(outcome.profileItemAcquisition);
+    if (auto* transaction = transaction_if<ProfileItemAcquisitionTransaction>(outcome)) {
+        const bool committed = state::commit_profile_item_acquisition(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=profile_acquire stage=transaction_commit result=ok"
                                    : "ev=profile_acquire stage=transaction_commit result=fail");
         return committed;
     }
-    if (outcome.hasItemDismantle) {
-        const bool committed = state::commit_item_dismantle(outcome.itemDismantle);
+    if (auto* transaction = transaction_if<ItemDismantleTransaction>(outcome)) {
+        const bool committed = state::commit_item_dismantle(transaction->pending);
         core::log::write(core::log::Channel::server,
                          committed ? core::log::Level::debug : core::log::Level::warn,
                          committed ? "ev=dismantle stage=transaction_commit result=ok"
